+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED
+
+ - WGPU renderer: Changed `flush_to_surface()` API to return a command buffer,
+   to let the application decide when to submit.
+
 ## [0.11.3] - 2024-12-26
 
  - WGPU renderer: Fix crash when rendering without always calling `set_size()`. (#226)